fix: remove custom domains from Vercel on downgrade to free
The customerSubscriptionDeleted webhook cleared page.customDomain in the
DB but left the domain attached to the Vercel project, so it kept
routing to the status page after downgrade. Collect the custom domains
of all workspace pages (including the extra pages the downgrade
deletes) inside the transaction and release them on Vercel after
commit, best-effort so a Vercel error doesn't fail the webhook.

Moves the Vercel domain helpers from the page router into
src/lib/vercel.ts so both routers share them.
